# kestra/openapi.yml
openapi: 3.0.1
info:
  title: Kestra
  license:
    name: Apache 2.0
    url: https://raw.githubusercontent.com/kestra-io/kestra/master/LICENSE
  version: 1.0.0
...
paths:
...
  /api/v1/users:
    get:
      tags:
      - Users
      summary: List users
      operationId: list_3
      responses:
        "200":
          description: list_3 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      tags:
      - Users
      summary: Create user
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserController.CreateUserRequest"
        required: true
      responses:
        "200":
          description: create_2 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/v1/users/{id}:
    get:
      tags:
      - Users
      summary: Get user by id
      operationId: get_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: get_2 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      tags:
      - Users
      summary: Delete user
      operationId: delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: delete_2 200 response
    patch:
      tags:
      - Users
      summary: Update user
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserController.UpdateUserRequest"
        required: true
      responses:
        "200":
          description: update_1 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/v1/users/{id}/reset-password:
    post:
      tags:
      - Users
      summary: Reset user password
      operationId: resetPassword
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserController.ResetPasswordRequest"
        required: true
      responses:
        "200":
          description: resetPassword 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      required:
      - createdAt
      - id
      - passwordHash
      - username
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        passwordHash:
          type: string
        salt:
          type: string
        disabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserController.CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        disabled:
          type: boolean
    UserController.ResetPasswordRequest:
      type: object
      properties:
        newPassword:
          type: string
    UserController.UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        disabled:
          type: boolean
...